# docs/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groundwork Query Runtime API",
    "version": "1.0.0",
...
    "license": { "name": "Apache-2.0", "url": "https://www.apache.org/licenses/LICENSE-2.0" }
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local quickstart (query-runtime, default QUERY_RUNTIME_ADDR :8080)" },
    { "url": "https://{host}", "variables": { "host": { "default": "gw.example.com", "description": "Managed Groundwork host" } }, "description": "Managed deployment" }
  ],
...
  "paths": {
...
    "/v1/leak-report": {
      "get": {
        "tags": ["Security"],
        "summary": "Scan connected external indexes for missing-scope documents",
        "operationId": "leakReport",
        "description": "Read-only compliance scan (e.g. over the connected GitHub org's indexed repos) that reports documents which should have been restricted - files pushed before an org was onboarded, excluded paths that got indexed, or permissions that were tightened after indexing.\n\n- Requires an API key with scope `audit`.\n- Returns `503 connector_unavailable` when no connector service is wired.",
        "responses": {
          "200": {
            "description": "Leak report scan completed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LeakResult" }
              }
            }
          },
          "401": {
            "description": "API key missing, invalid, or expired.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Key lacks scope `audit` (`insufficient_scope`).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "503": {
            "description": "Connector service not wired (`connector_unavailable`).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Groundwork-API-Key",
        "description": "Tenant API key. Scoped; `admin` satisfies any scope. Also accepted as `Authorization: Bearer <key>`."
      },
      "identityAssertion": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Groundwork-User-Assertion",
        "description": "Verified end-user identity assertion (JWT, signed via the configured HMAC secret or OIDC issuer). Required by endpoints that must attribute actions to a human operator."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Error envelope: a stable machine-readable `error` code.",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "insufficient_scope" }
        }
      },
...
      "LeakResult": {
        "type": "object",
        "required": ["findings"],
        "properties": {
          "findings": {
            "type": "array",
            "description": "Documents that should have been restricted but are reachable in an indexed external system.",
            "items": { "$ref": "#/components/schemas/LeakFinding" }
          }
        }
      },
      "LeakFinding": {
        "type": "object",
        "required": ["kind", "severity", "title", "detail"],
        "properties": {
          "kind": { "type": "string", "example": "missing_org_scope" },
          "severity": { "type": "string", "enum": ["high", "medium", "low"], "example": "high" },
          "title": { "type": "string", "example": "Pre-onboarding commit visible to everyone" },
          "detail": { "type": "string" }
        }
      },
...
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}